+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 
 - Bump to opentelemetry-proto v0.19.0. (#5823)
 - Expose `Scope.Attributes` in pdata (#5826)
+- Add support to handle 404, 405 http error code as permanent errors in OTLP exporter (#5827)
 
 ### 🧰 Bug fixes 🧰